pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devgen"
version = "0.1.0"
description = "A collection of developer tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [
{ name = "Sankalp Tharu", email = "sankalptharu50028@gmail.com" },
]
license = { text = "GPL-3.0-or-later" }
keywords = ["devgen", "cli", "git", "changelog", "gitignore", "license", "commit-generator", "commit"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
    "anthropic>=0.75.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "openai>=2.11.0",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "toml>=0.10.2",
    "typer>=0.20.0",
]
[project.urls]
Homepage = "https://github.com/S4NKALP/devgen"

[project.scripts]
devgen = "devgen.cli.main:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["devgen*"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.9",
]
